backend/requirements.txt
# Arduino Flash Agent Dependencies
websockets>=10.0
pyserial>=3.5